Move fetch dependents action to shared folder
