+++ src/applications/dependents/dependents-verification/reducers/index.js
@@ -1,7 +1,7 @@
 import { createSaveInProgressFormReducer } from 'platform/forms/save-in-progress/reducers';
 
 import formConfig from '../config/form';
-import dependents from './dependents';
+import dependents from '../../shared/reducers/dependents';
 
 export default {
   form: createSaveInProgressFormReducer(formConfig),
